requirements.txt
Flask==3.0.3
pytest==8.3.3
pytest-mock==3.14.0
pypdf==5.1.0
python-docx==0.8.11
openpyxl==3.1.2
Werkzeug==3.0.1
python-pptx==1.0.2
easyocr==1.7.2
scikit-learn==1.4.1
joblib==1.3.2
numpy==1.26.4
Pillow==10.2.0
